backend/requirements.txt
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.10
fastapi>=0.68.0
uvicorn>=0.15.0
fastapi[all]
passlib[bcrypt]
requests>=2.26.0
python-dotenv>=0.19.0
